% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline}
\alias{baseline}
\title{Baseline Data for RHISEA}
\format{
A data frame (ou list) with...
}
\source{
Your Lab / Gulf Fisheries Centre
}
\usage{
baseline
}
\description{
A dataset containing baseline stock characteristics.
}
\keyword{datasets}
