% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final.R
\name{.resample_baseline_data_helper}
\alias{.resample_baseline_data_helper}
\title{Helper: resample baseline data per stock}
\usage{
.resample_baseline_data_helper(
  original_baseline_list,
  resampled_sizes,
  stock_names_for_error,
  nv_fallback
)
}
\arguments{
\item{original_baseline_list}{List of baseline matrices.}

\item{resampled_sizes}{Integer vector of sizes for resampling.}

\item{stock_names_for_error}{Character vector of stock names for error messages.}

\item{nv_fallback}{Integer. Number of variables to use if a matrix is empty.}
}
\value{
A list of resampled matrices.
}
\description{
Helper: resample baseline data per stock
}
\keyword{internal}
