% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_millar_theta4}
\alias{estimate_millar_theta4}
\title{Estimate Millar Theta4 Parameter}
\usage{
estimate_millar_theta4(
  class_predictions_mixed_sample,
  PHI_matrix,
  np,
  use_accelerated_em = TRUE,
  ...
)
}
\arguments{
\item{class_predictions_mixed_sample}{Numeric vector of predicted classes for the mixed sample}

\item{PHI_matrix}{Confusion matrix used for correction}

\item{np}{Integer number of populations}

\item{use_accelerated_em}{Logical; whether to use the accelerated EM version}

\item{...}{Additional arguments passed to internal functions}
}
\value{
Numeric value of the estimated theta4 parameter
}
\description{
This function estimates the theta4 parameter using the Millar method.
}
\examples{
\dontrun{
pred_classes <- c(1, 2, 1, 2)
phi_mat <- matrix(c(0.8, 0.2, 0.2, 0.8), 2, 2)
theta4 <- estimate_millar_theta4(pred_classes, phi_mat, np = 2)
}
}
