% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixture}
\alias{mixture}
\title{Mixture dataset for RHISEA examples}
\format{
A data frame with X rows and 2 variables:
\describe{
  \item{d13c_ukn}{Carbon isotope ratio for unknown samples}
  \item{d18o_ukn}{Oxygen isotope ratio for unknown samples}
}
}
\usage{
data(mixture)
}
\description{
A synthetic dataset containing isotope values for unknown individuals.
}
\keyword{datasets}
