% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ordvec}
\alias{ordvec}
\title{Generate Ordered Random Values (Order Statistics based)}
\usage{
ordvec(num, upper)
}
\arguments{
\item{num}{Integer, the number of ordered values to generate.}

\item{upper}{Numeric, the maximum value for the scaled output. The generated values
will be in the approximate range (0, \code{upper}].}
}
\value{
A numeric vector of length \code{num}, containing ordered values scaled
approximately between 0 and \code{upper}.
}
\description{
Produces an ordered vector of \code{num} random values, scaled to be approximately
between 0 and \code{upper}. This function can be used as a basis for certain types
of random sampling schemes. The logic is similar to generating order statistics
from an exponential distribution.
}
\examples{
ordvec(5, 100) # Generates 5 ordered numbers, max value around 100
ordvec(10, 10) # Generates 10 ordered numbers, max value around 10
}
