% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldf.R
\name{get_phi}
\alias{get_phi}
\title{Get Phi Matrix}
\usage{
get_phi(predicted_class, true_labels, np)
}
\arguments{
\item{predicted_class}{Integer vector of predictions}

\item{true_labels}{Factor of true labels}

\item{np}{Number of populations (levels)}
}
\value{
Matrix of phi values
}
\description{
Calculate the phi matrix for stock composition analysis.
}
\examples{
\dontrun{
pred <- c(1, 2, 1, 2)
true <- c(1, 2, 1, 1)
phi_matrix <- get_phi(pred, true, np = 2)
}
}
