% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_hisea_theta}
\alias{plot_hisea_theta}
\title{Plot HISEA theta estimates with error bars
utils::globalVariables(c("Stock", "Mean", "Estimator", "SD", "ActualValue"))}
\usage{
plot_hisea_theta(summary_report, stock_names = NULL, actual_proportions = NULL)
}
\arguments{
\item{summary_report}{Output from code{create_hisea_summary_report}.}

\item{stock_names}{Optional character vector for stock labels.}

\item{actual_proportions}{Optional numeric vector of true proportions.}
}
\value{
A ggplot object.
}
\description{
Plot HISEA theta estimates with error bars
utils::globalVariables(c("Stock", "Mean", "Estimator", "SD", "ActualValue"))
}
