% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{print_hisea_summary}
\alias{print_hisea_summary}
\title{Print HISEA Summary Report to Console and/or File}
\usage{
print_hisea_summary(
  summary_report,
  nsamps,
  actual_proportions,
  means_vars = NULL,
  sds_vars = NULL,
  standard_sizes = NULL,
  stocks_names,
  output_file = "hisea_summary_report.txt",
  header = NULL,
  run_type = "SIMULATION",
  nv_used = NULL,
  seed_val = NULL,
  nmix_simulated = NULL
)
}
\arguments{
\item{summary_report}{Output from code{create_hisea_summary_report}.}

\item{nsamps}{Number of simulation/bootstrap replicates.}

\item{actual_proportions}{True composition vector.}

\item{means_vars}{Matrix of mean variable values.}

\item{sds_vars}{Matrix of SD of variable values.}

\item{standard_sizes}{Vector of baseline sample sizes.}

\item{stocks_names}{Character vector of stock names.}

\item{output_file}{Path to the output text file.}

\item{header}{Optional custom header line.}

\item{run_type}{Character string of run type.}

\item{nv_used}{Integer, number of variables.}

\item{seed_val}{Integer, seed value.}

\item{nmix_simulated}{Integer, size of simulated mixture.}
}
\description{
Print HISEA Summary Report to Console and/or File
}
