% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final.R
\name{process_hisea_input}
\alias{process_hisea_input}
\title{Process HISEA Input Data}
\usage{
process_hisea_input(
  input,
  type = c("std", "mix"),
  nv = NULL,
  stock_col = NULL,
  var_cols_mix = NULL,
  var_cols_std = NULL
)
}
\arguments{
\item{input}{Input data (file path or data frame)}

\item{type}{Type of input ("std" or "mix")}

\item{nv}{Number of variables}

\item{stock_col}{Stock column name}

\item{var_cols_mix}{Variables for mixture data}

\item{var_cols_std}{Variables for standard data}

\item{out_path}{Output path}

\item{stocks_names}{Stock names}
}
\value{
Processed data
}
\description{
Process HISEA Input Data
}
\keyword{internal}
