% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final.R
\name{write_mix_from_dataframe}
\alias{write_mix_from_dataframe}
\title{Write Mixture Data to File}
\usage{
write_mix_from_dataframe(df, var_cols_mix, file_path = "hisea.mix")
}
\arguments{
\item{df}{Data frame}

\item{var_cols_mix}{Variables to write}

\item{file_path}{Output file path}
}
\value{
None
}
\description{
Write Mixture Data to File
}
\keyword{internal}
