% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final.R
\name{write_std_from_dataframe}
\alias{write_std_from_dataframe}
\title{Write Standard Data to File}
\usage{
write_std_from_dataframe(df, stock_col, var_cols_std, file_path = "hisea.std")
}
\arguments{
\item{df}{Data frame}

\item{stock_col}{Stock column}

\item{var_cols_std}{Variables to write}

\item{file_path}{Output file path}
}
\value{
None
}
\description{
Write Standard Data to File
}
\keyword{internal}
